data_pipeline/requirements.txt
python-dotenv
psycopg2-binary
playwright
unidecode